Fixed: change typing on register_pattern
Previously the `register_pattern` method had a type declaration of `Pattern` for the `pattern` argument, but the instance checker would raise an error if the pattern was not a string.

This commit changes the type declaration to accept `str` rather than Pattern. Changing the type annotation was chosen as it was the least disruptive to the current behaviour of the library.

A small test was added to check that this function does not accept compiled regex statements.

Fixes #59