Don't try empty path if '/' was not found.
`Requested URL  not found` when accessing `/` without `strict_slashes` enabled. I did not check whether there is a similar problem with subdirectories too. The URL requested by the client should be shown, not the one with a slash removed.

The message is also fixable in sanic/router.py:42 by using `path` instead of `e.path`.